// src/components/product/types/IProductPreviewSettings.ts
export type IProductPreviewSettings = {
	selectorImage: string;
	selectorCategory: string;
	selectorTitle: string;
	selectorDescription: string;
	selectorButton: string;
	selectorPrice: string;
	classCategorySoft: string;
	classCategoryHard: string;
	classCategoryOther: string;
	classCategoryAdditional: string;
	classCategoryButton: string;
};